master: add text editor

+++ back/src/Controller/BlogController.php
@@ -13,24 +13,27 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Component\String\Slugger\SluggerInterface;
+use Knp\Component\Pager\PaginatorInterface;
 /**
- * @Route("/blog")
  * @IsGranted("ROLE_USER")
  */
 class BlogController extends AbstractController
 {
     /**
-     * @Route("/", name="blog_index", methods={"GET"})
+     * @Route("/blogs/{page}", defaults={"page"=1}, name="blog_index", methods={"GET"})
      */
-    public function index(BlogPostRepository $blogPostRepository): Response
+    public function index(BlogPostRepository $blogPostRepository, PaginatorInterface $paginator, int $page): Response
     {
+        $_blogPosts = $this->getDoctrine()->getRepository(BlogPost::class)->findBy(array(), array('id' => 'DESC'));
+        $_pageBlogPosts = $paginator->paginate($_blogPosts, $page, 10);
+
         return $this->render('blog/index.html.twig', [
-            'blog_posts' => $blogPostRepository->findAll(),
+            'blogPosts' => $_pageBlogPosts,
         ]);
     }
 
     /**
-     * @Route("/user/{id}", name="blog_posts_user", methods={"GET"})
+     * @Route("/blog/user/{id}", name="blog_posts_user", methods={"GET"})
      */
     public function user(BlogPostRepository $blogPostRepository, User $user): Response
     {
@@ -42,7 +45,7 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/manage", name="blog_user_management", methods={"GET"})
+     * @Route("/blog/manage", name="blog_user_management", methods={"GET"})
      */
     public function manage(BlogPostRepository $blogPostRepository): Response
     {
@@ -54,7 +57,7 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/new", name="blog_post_new", methods={"GET","POST"})
+     * @Route("/blog/new", name="blog_post_new", methods={"GET","POST"})
      */
     public function new(Request $request, SluggerInterface $slugger): Response
     {
@@ -83,6 +86,7 @@ class BlogController extends AbstractController
 
                 $blogPost->setPicture($newFileName);
             }
+            $blogPost->setAuthor($this->getUser());
 
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($blogPost);
@@ -98,7 +102,7 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="blog_post_show", methods={"GET"})
+     * @Route("/blog/{id}", name="blog_post_show", methods={"GET"})
      */
     public function show(BlogPost $blogPost): Response
     {
@@ -108,10 +112,14 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/{id}/edit", name="blog_post_edit", methods={"GET","POST"})
+     * @Route("/blog/{id}/edit", name="blog_post_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, BlogPost $blogPost): Response
     {
+        if($this->getUser()->getId() != $blogPost->getAuthor()->getId()) {
+            return $this->redirectToRoute('blog_index');
+        }
+
         $form = $this->createForm(BlogPostType::class, $blogPost);
         $form->handleRequest($request);
 
@@ -128,10 +136,14 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="blog_post_delete", methods={"DELETE"})
+     * @Route("/blog/delete/{id}", name="blog_post_delete", methods={"DELETE"})
      */
     public function delete(Request $request, BlogPost $blogPost): Response
     {
+        if($this->getUser()->getId() != $blogPost->getAuthor()->getId()) {
+            return $this->redirectToRoute('blog_index');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$blogPost->getId(), $request->request->get('_token')) && $this->getUser() == $blogPost->getAuthor()) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($blogPost);
